A log-likelihood calculation time comparison for different numbers of traits and option-sets

A log-likelihood calculation time comparison for different numbers of traits and option-sets

BenchmarkRvsCpp(ks = c(1, 2, 4, 8), includeR = TRUE,
  includeTransformationTime = TRUE, optionSets = NULL,
  includeParallelMode = TRUE, doProf = FALSE,
  RprofR.out = "RprofR.out", RprofCpp.out = "RprofCpp.out",
  verbose = FALSE)

Arguments

ks
a vector of positive integers, denoting different numbers of traits. Default: c(1, 2, 4, 8).

includeR
logical (default TRUE) indicating if likelihood calculations in R should be included in the benchmark (can be slow).

includeTransformationTime
logical (default TRUE) indicating if the time for PCMApplyTransformation should be included in the benchmark.

optionSets
a named list of lists of PCM-options. If NULL (the default) the option set is set to DefaultBenchmarkOptions(k, includeParallelMode) for each k in ks (see the code in PCMBaseCpp:::DefaultBenchmarkOptions).

includeParallelMode
logical (default TRUE) indicating if the default optionSet should include parallel execution modes, i.e. setting the option PCMBase.Lmr.mode to 21 instead of 11. This argument is taken into account only with the argument optionSets set to NULL (the default).

doProf
logical indicating if profiling should be activated (see Rprof from the utils R-package). Default: FALSE. Additional arguments to Rprof can be specified by assigning lists of arguments to the options 'PCMBaseCpp.ArgsRprofR' and 'PCMBaseCpp.ArgsRprofCpp'. The default values for both options is list(append = TRUE, line.profiling = TRUE).

RprofR.out
character strings indicating Rprof.out files for the R and Cpp implementations; ignored if doProf is FALSE. Default values: 'RprofR.out' and 'Rprofcpp.out'.

RprofCpp.out
character strings indicating Rprof.out files for the R and Cpp implementations; ignored if doProf is FALSE. Default values: 'RprofR.out' and 'Rprofcpp.out'.

verbose
logical indicating if log-messages should be printed to the console during the benchmark. Default FALSE.

Value

a data.table for results similar to the data.table returned from MiniBenchmarkRvsCpp with additional columns for k, option-set and the type of model.