A S3 generic for creating C++ backend objects given a model, data and a tree.

Replace calls to PCMInfo() with this method in order to use C++ for likelihood calculation.

PCMInfoCpp(X, tree, model, metaI = PCMInfo(X, tree, model, verbose,
  preorder = PCMTreePreorderCpp(tree)), verbose = FALSE, ...)

Arguments

X
a k x N numerical matrix with possible NA and NaN entries. Each column of X contains the measured trait values for one species (tip in tree). Missing values can be either not-available (NA) or not existing (NaN). Thse two values have are treated differently when calculating likelihoods: see PCMPresentCoordinates.

tree
a phylo object with N tips.

model
an S3 object specifying both, the model type (class, e.g. "OU") as well as the concrete model parameter values at which the likelihood is to be calculated (see also Details).

metaI
a list returned from a call to PCMInfo(X, tree, model), containing meta-data such as N, M and k. Default: PCMInfo(X, tree, model, verbose, preorder=PCMTreePreorderCpp(tree))

verbose
logical indicating if some debug-messages should be printed. Default: FALSE

...
passed to methods.

Value

a list to be passed to PCMLik as argument metaI.